research_agent/pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Research agent with DuckDuckGo search tool"
requires-python = ">=3.10"
dependencies = [
    "strands-agents[openai]>=1.0.0",
    "ddgs>=9.0.0",
    "bedrock-agentcore>=1.0.0",
]